+++ views/recommendationList.handlebars
@@ -19,11 +19,8 @@
             title="addFriendIconButton"
             id="addFriendIconButton"
             class="fa-solid fa-user-group rounded-pill border-0 p-3"
-          >
+          >+
           </button>
-          {{! <script>
-            console.log("bad");
-          </script> }}
         </form>
       {{/if}}
     </div>
